backend/requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
sqlparse==0.4.4
psycopg2-binary==2.9.7
pytz==2021.1
PyJWT==2.1.0
python-dotenv==0.19.2
